修改 收藏歌曲到歌单只显示创建的歌单

+++ Shared/State/AppState.swift
@@ -113,7 +113,7 @@ extension AppState {
         //喜欢的音乐ID
         var likedIds = [Int64]()
         //喜欢的音乐歌单ID
-        var likedPlaylistId: Int = 0
+        var likedPlaylistId: Int64 = 0
         var userPlaylistIds = [Int64]()
     }
